Add new feature to sim, to help with debug.
The first 16 break_ codes (0-15) are interpreted as a "debug marker"
which prints all register content to the screen, and continues (does
not stop execution). It also prints the current instruction count,
which can be used with --stop-sim-at flag on command line, once
you do figure out where you want to stop the sim.

The intent is to be able to place these visible markers in code,
and do an abbreviated trace.