Fix unary-op bit not. In the path where it creates a heap number for the result, it returned it in a0, needs to be v0.
Minor other optimization to make overflow check a bit faster.

The msjsunit test bit-not.js takes 2m 17s to complete on my macbook
under simulator, but it does complete without error.

It runs a huge loop waiting for GC to bug. This means that with normal
timeout settings, we will fail this test...