Use FP trunc.w.d or trunc.l.d to convert double to int or long-int.
Fixes sunspider string-tagcloud.js and v8 v4 benchmark v8-raytrace.js

The  cvt_w_d (or cvt_l_d) uses FPU "current rounding mode" which
defaults to "round-towards-nearest". C code usually assumes round-toward-
zero", which is what trunc does.

To make matters worse, the simulator does not emulate rounding mode, it
just truncated, so we had code that would work on sim, but not on
hardware.

(Note that this does NOT change the one cvt_w_d() in MOD workaround
in HandleBinaryOpSlowCases() since Nemanja has been working on that
code. We will re-visit the need for that after the merge is complete.)