Add FP trunc instructions.
We have been using cvt.w.d, cvt.l.d to convert from FP to integer.
This instruction is subject to the rounding modes as specified in FCCR, which default
to round-nearest, which is not what we want.

Rather than screwing around with rounding modes, I am adding in the TRUNC instruction
which rounds towards zero, which is what gcc uses for mips hard-float.

This commit just addes the instructions to asm, disasm, and sim. A later commit
will change generated code to use the new instructions.