Revert "Fix for sunspider string-tagcloud.js"
This reverts commit 5dde62112e1949e01ef5c23e55475f84b6eaf599.
This was a false fix, it did make the test pass, but it was not the
root cause. It also was incorrect, and therefore caused problems in
other code.

The real issue was using the incorrrect FP rounding mode with
cvt.w.d. The correct fix will be committed shortly.