Version 2.3.3
Fixed error when building the d8 shell in a fresh checkout.

Implemented Function.prototype.bind (ES5 15.3.4.5).

Fixed an error in inlined stores on ia32.

Fixed an error when setting a breakpoint at the end of a function that does not end with a newline character.

Performance improvements on all platforms.


git-svn-id: https://v8.googlecode.com/svn/trunk@5133 ce2b1a6d-e550-0410-aec6-3dcde31c8c00