Version 2.4.7
Changed the command-line flag --max-new-space-size to be in kB and the flag --max-old-space-size to be in MB (previously they were in bytes).

Added Debug::CancelDebugBreak to the debugger API.

Fixed a bug in getters for negative numeric property names (https://bugs.webkit.org/show_bug.cgi?id=46689).

Performance improvements on all platforms.

git-svn-id: https://v8.googlecode.com/svn/trunk@5563 ce2b1a6d-e550-0410-aec6-3dcde31c8c00