[MIPS merge fixup] for Change the LoadIC calling convention ....
Change the LoadIC calling convention so that the receiver is both on top
of the stack and also in r0. This makes sense because the receiver is usually
in r0 anyway. We may remove it from the stack later. Also removes some spilled
scopes from the code generator allowing it to keep expression temporaries
in registers more.