[MIPS merge fixup] hack Reference::GetValue() to work
This adds some changes that were in earlier google commits that have
not yet been merged for mips. All related to the changing IC and
virtual-frame stuff.

This is likely to cause merge conflicts for issue 60. But it is needed for
code I am working on for issue 66. Trying to put this code in the
same state it was in prior to the original commit of 1fd6a2e76,
"ARM: Don't require the receiver on the stack for load IC".