[MIPS merge] Support multi-register Push macro.
This is a port to Mips of the Arm change 485c15bfa,
"Simplify the use of the stm instruction on ARM".

This supports pushing registers in _any_ order, not the
in-flexible order of MultiPush() and MultiPushReversed().

Most usages of those functions have been replaced, however,
the functions remain. Also note that there is no multi-register
version of Pop(), but there probably should be.

This commit does not fix the Push() usage in stub-cache-mips.cc,
which I will update next.