[MIPS merge] More precise break points and stepping when debugging.
Port 3baa5c30 to mips.

Note that Mips support is incomplete/incorrect, since we do not yet
support FULL codegen. We are still limping along trying to debug in the
optimizing codegen, which is known not to work, and we have not
fixed all debugger tests, even when that mode is supported.

A new issue 77 will be raised, to document the details.