[MIPS merge] ARM: Don't emit a write barrier for an inlined keyed load if ...
Note this is actually for inlined keyed STORE (orig comment incorrect)

ARM: Don't emit a write barrier for an inlined keyed load if the right hand
side is a literal like true, false, etc. Also if the value is not a likely
Smi we inline the newspace check.

Port c2dd344c9 to mips.