[MIPS merge] Simplify the transitions in the Binary Op ICs.
Simplify the transitions in the Binary Op ICs.  Now a single call
to the runtime will both patch in the more specialized binary op
stub and calculate the answer.  This eliminates the need to call
both the rest of the binary op and the patching runtime call.  The
runtime routines are altered to be more agressive in returning
Smis so we don't get spurious heap numbers as inputs to binary ops
while we are patching the binary op ICs.

Port 987ad81a to mips.