[MIPS] Fix BitCast usage in simulator-mips.cc
Google commit 04e4bae4 removed the -fno-strict-aliasing flag, and fixed
some usages of BitCast. This provoked a failure in gcc 4.4.3.

For the external calls to C code in the Simulator, I had to change the
BitCast that converted a double return value to two 32-bit integers in
registers v0,v1. Could not make BitCast work on the 64-bit source reg.

So I changed it to use memcpy() directly, as is used in simulator-arm.cc