[MIPS merge] backend opt for ToBoolean: JIT code generation for ToBool
Port arm fix e66dfc49 to mips. Here is the Arm commit message:

ARM: backend opt for ToBoolean: JIT code generation for ToBool

Upgraded the CodeGenerator::ToBoolean() function in the ARM
backend to use complete JIT code generation and not make runtime
calls to ToBool (when VFP is enabled).

This change also includes the vcmp VFP instruction that supports
a constant 0.0 as the second operand.