[fcg] Workaround to fix the indexed-accessors.js bug.
This is NOT a long-term fix. But it does seem to fix this issue.
I still believe that there is some mistake in nesting the contexts
somewhere, since Arm code does not need a fix like this, to recognize
the correct context.

Without this woraround, the bug can be seen with this fragment:

a = function goo() {};
if (a |= '') {};
print(a);

which _should_ print "0", but incorrectly prints "function goo() {}".

The reason is that assignment a |= '' , when it runs
FullCodeGenerator::VisitAssignment() should be of AccumulatorValueContext,
but is of TestContext(). Therefore the 'context()->Plug(v0)' in
EmitBinaryOp() converts the result to Boolean, and does not store it.
The results (0) is then lost. By changing it to AccumulatorValueContext,
it will get saved via a StoreIC.