[MIPS] Fix cctest test-api/CallAsFunction.
Fix a bug in Builtins::Generate_FunctionCall(), where we wrote the function
back to the stack, just past comment 3b. The actual bug fix is just this
small diff, where the saved value in t1 no longer pointed to correct
stack location:

   // Restore the function in case it has been modified.
-  __ sw(a1, MemOperand(t1));
+  __ sll(at, a0, kPointerSizeLog2);
+  __ addu(a2, sp, at);
+  __ sw(a1, MemOperand(a2, -kPointerSize));

The rest of the function was refactored a bit, since it saved/restored 
extra stuff to the stack, which I did not think was necessary. There
are a few more instructions now, and a few less memory accesses.