[MIPS] Fix for the fast case of GenerateMathPow()
There were three bugs in our implementation of that function. One of them
is inherited from ARM version of the same function, which causes v8
unexpectedly to fall-back to runtime when computing power of a number
(this also explains why ARM did not achieved significant performance
increase).

There is still an open issue regaring the use of OBJECT_NOT_SMI flag
to the ObjectToDoubleFPURegister() macro-assember routine. Documented
still in gh issue 78.