[MIPS] Fix debugger source-position problems.
In EmitNamedPropertyAssignment, mips code was not storing the source
position RelocInfo at the correct time, resulting in the pc() of the
position getting stored _after_ that for subsequent code_target:
STORE_IC stub (in this case).

The solution is to put an explicit call to WriteRecordedPositions().

I am adding this to other calls to CodeForSourcePosition(), since a
similar issue seems likely, but I do not have test results to prove that.

At this time, I am not adding this change to full-codegen-mips.cc,
but I suspect it might be needed.

This change fixes:
     cctest/DebugStepLinear
     mjsunit/debug-breakpoints.js