[MIPS] Fix debugger source-position for  full-codegen.
As in recent change to codegen-mips.cc, the source position RelocInfo
is getting associated with the wrong pc value, and comes after a
nearby code_target. This breaks the interators used to search for
positions and code_targets. This behavior appears to be arch-specific
for mips.

Fix is to explicitly write RelocInfo containing source position before
we reach the code_target.

This change fixes:  (with --always-full-compiler)
     cctest/DebugStepLinear
     mjsunit/debug-breakpoints.js