[MIPS] Fix the source-position issue, correctly this time.
I had a previous fix that made calls to WriteRecordedPositions() all over
the code generators. Pretty damn ugly ... but worked.

Now I understand the situation better, and have a clean fix in
MacroAssembler::Call().

The problem was that for code targets, a Call() becomes:
    li(t9, target)
    jalr(t9)

And the assembler jalr() already wrote recorded source pos.

The problem is that we must save previous source-line positions before
the li(), since that li macro-instruction also records reloc info. When it
does that it updates the pc_ value associated with the reloc info, and then
previous source positions have incorrect pc values.

So putting an explicit WriteRecordedPositions() before the li() in Call()
fixes the issue.