[MIPS] Temporary workaround, must be reverted. Remove TypeRecordingBinaryOpStub.
This stub was used in one place, FullCodeGenerator::EmitBinaryOp.

It caused several mjs & moz tests to fail. There are comments that the
patching is not correctly implemented, that could be a reason for the fail.

In full-codegen-arm, this stub is used in 4 places. Here we had used it
in just one. Maybe we get an inconsistent view of Smi types since it
is not used everywhere.

Regarless, it seems that testing mjust unit with --always-full-compiler
is a good test of the TypeRecordingBinaryOpStub.

Before hack: mjsunit  [02:18|%100|+689|-10]:Done
After  hack: mjsunit  [02:18|%100|+696|-3]:Done