[MIPS] Update CEntryStub return address calculations.
Add some debug code for computing ra that had been missing,
for debug mode.

Assert-check fixed instruction count sequences, which are needed
for mips.

It looks like a trampoline-pool block is needed, but surprisingly,
there is not an equivalent contant-pool block on arm. It may be that
the exact 'ra' value is not critically important, since it is not
used for an actual return.