[MIPS] Simplify li psuedo-instruction in the 2-instruction case.
The case for using 2 instructions is for adresses, so the optimization
of trying to use one instruction (loading hi or lo 16 bits) + nop is
almost never going to work. Yet we pay cycles every time we patch one
of these.

Doug tried this, and showed that it measurably improves benchmark scores.

Updates MacroAssembler::li() and Assembler::set_target_address_at()