[MIPS] Standardize on lower case push() and pop() for single argument.
Currently we have both push() and Push(), and pop() and Pop(). Deprecate the
uppercase ones.

History is that Mips created Push() and Pop() as macro-instructions, since
they emit two instructions each.

But the arch-independent code expects to call push() and pop() so aliases
were made. Then much code was ported in and left the lower-case push() and
pop() all over. So we have inconsistent usage, which is bad.

Since we are required to have push()/pop() for arch-indep code, I am removing
Push()/Pop(). Of course that violates other conventions, like macros-assembler
instuctions starting with capital, but we have other examples, like li(), so
we will let that rule slide in favor of consistency.

Note that this does NOT change names of the 2, 3, or 4-argument Push()/Pop()
instances. However, I am removing some partly-implemented conditional
arguments on those, which I mistakenly copied from the arm versions.