[MIPS] Allow patching JAL->JALR in functions moved by GC.
The JALR->JAL optimization had a fault, found on Android, but not
on Linux. If code that has been patched to JAL gets moved by gc to
a different 256 MB page, then the JAL must be reverted to JALR.

This did not happen before, now we have used RelocInfo::apply(),
which is called from Code::Relocate(), to make this change.

All comments apply also to J->JR.