[MIPS crankshaft] Cleaned up & finalized deoptimizer-mips.cc.
Fixed all presubmit errors except one that deliberately marks a potential bug that needs to be revisited later (when the problem it causes becomes more apparent).
There are still some TODOs in the file that mark things that need to be checked or optimized, but these should be working in their current form as well.