[MIPS crankshaft] Print read/write to bad address in Simulator.
This is an addition to Daniel's print of read bad addr from 1672436ff.
I added print of a bad write, as well as added the PC address.

The problem he found was that the access was in the delay slot of a
branch instruction, and so the normal sim output display was
confusing, since it showed the branch, not the delay-slot instruction.

I hope addition of PC is helpful in this case.