[MIPS crankshaft] Fix rare bug in LCodeGen::EmitTypeofIs()
In some cases, a couple mozilla tests would fail when not matching
any of the if () clauses, and falling to the final else, Branch(false_label).

In that path, the cmp1, cmp2 Register references had been left at no_reg,
which triggered an assertion.

It is not clear to me why this path is only hit some of the time, not every
time....

Here are two tests that did periodically fail:

/opt/paul/rb-v8/shell_g --enable-slow-asserts --debug-code --verify-heap --expose-gc /opt/paul/rb-v8/test/mozilla/mozilla-shell-emulation.js /opt/paul/rb-v8/test/mozilla/data/shell.js /opt/paul/rb-v8/test/mozilla/data/ecma_3/shell.js /opt/paul/rb-v8/test/mozilla/data/ecma_3/RegExp/shell.js /opt/paul/rb-v8/test/mozilla/data/ecma_3/RegExp/15.10.4.1-3.js

/opt/paul/rb-v8/shell_g --enable-slow-asserts --debug-code --verify-heap --expose-gc /opt/paul/rb-v8/test/mozilla/mozilla-shell-emulation.js /opt/paul/rb-v8/test/mozilla/data/shell.js /opt/paul/rb-v8/test/mozilla/data/js1_5/shell.js /opt/paul/rb-v8/test/mozilla/data/js1_5/Regress/shell.js /opt/paul/rb-v8/test/mozilla/data/js1_5/Regress/regress-465137.js