[MIPS crankshaft] Code style changes in LCodeGen::EmitTypeofIs.
Small stuff on style. No functional change.

When wrapping too many parameters in a function call, the left side
must be aligned with a opening parenthesis in line above.

(A different case is  wrapping a long statement, which the preferred
method is to do a 4-space indent.)

My own preference (not google's) is when wrapping one of our Branch()
macro-instructions, is to keep the condition code (cc) and the two
reg operands on the same line. Hence, I wrapped one like this:

  __ Branch(USE_DELAY_SLOT, false_label,
            gt, scratch, Operand(LAST_NONCALLABLE_SPEC_OBJECT_TYPE));

But if it doesn't fit due to length, don't worry about it.