[MIPS crankshaft] Added BranchF (floating-point branch) macro, implemented lithium functions using it.
BranchF is a little different than the regular Branch macro because of MIPS's special way of doing FPU comparison and jumps.
Also, this should be considered somewhat experimental, there are some cases that are probably not tested in the code, especially regarding NaN values.
Some code stubs have also been updated to use this new macro.