[MIPS crankshaft] Implemented LCodeGen::EmitNumberUntagD and LCodeGen::DoNumberUntagD.
This allows mjsunit div-mod.js to be compiled but exposes two bugs.
One is a double comparison problem that probably originates from the BranchF commit.
The other is a random occurence where the system occasionally tries to save odd numbered double registers.
Also removed a delay slot usage from one of the BranchF instructions because the simulator had some problem with it.