[MIPS crankshaft] Two small fixes related to the bugs found while running div-mod.js
Fixed an invalid assertion that wouldn't allow f0 to be saved in safepoints.
Added 'al' and unsigned comparison cases to BranchF. I'm not sure if this is the proper way to handle them (why are unsigned comparisons used on doubles?) but it seems to be working.

div-mod.js now passes.

Also removed an abort based on an invalid assumption in EmitBranchF.