[MIPS crankshaft] ToBooleanStub now also returns the result in v0.
Intended or not, ToBooleanStub caused an error in crankshaft code which expected it to return the value in v0 (but instead returned it in a tos register, a0 in most cases).
This could have been fixed by a much smaller change in lithium-codegen-mips.cc, however this way we avoid this kind of mistake in the future. v0 is set up in the delay slot right after Ret so there should be zero change in performance.

This fixes several mjsunit and mozilla failures ("Cannot read property 'length' of null"-type errors).