[MIPS crankshaft] Fix DoAccessArgumentsAt().
There was a missing sub of index from length, and I also think (but not quite
certain) that we cannot overwrite the 'arguments' register. So I used 'at' as
a temporary.

This function can be optimized by combining the subu(length, length, index),
and the DeoptimizeIf() as they do in Arm code, but I can't quite get the
solution tonight.