[MIPS crankshaft] Slightly re-worked FullCodeGenerator::EmitStackCheck.
Instead of using MacroAssembler's Branch the Branch sequence is now generated manually to make sure it's constant.

Also in the comments:
Some new Branch and trampoline pool changes can cause Branch to emit a skip-or-jump sequence instead of the normal Branch. It seems that the "skip" part of that sequence is about as long as this Branch would be so it is safe to ignore that. Still, this may cause problems.

This fixes mozilla regress-80981.