[MIPS crankshaft] Implemented LCodeGen::DoInstanceOfKnownGlobal,  LCodeGen::DoDeferredLInstanceOfKnownGlobal.
Also implemented the crankshaft-specific parts of InstanceofStub::Generate and MacroAssembler::PatchRelocatedValue.
This is a little different from ARM: they rewrite the value in the const pool while on MIPS the lui/ori immediates are rewritten instead.
This utilizes a new MacroAssembler::FlushICache function that calls CPU::FlushICache from generated code.