[MIPS crankshaft] Fix a BinaryOpStub marker bug.
This fixes a failure while running the negate.js test from mjsunit.
In crankshaft mode, a call to BinaryOpStub is not followed by a marker,
yet in PatchInlinedSmiCode(), the existence of a marker is checked in
full-codegen and crankshaft modes. Therefore if the marker is a normally
used instruction, it could lead to a false detection of the marker and
cause the program to access random locations in memory.

Note that google just commited 14752c97 to add a nop as the marker
in other arch. With this patch, we do not nbeed that nop.