[MIPS crankshaft] Cleaned up two TODOs in LCodeGen::DoDeferredTaggedToI.
Added an assertion to make sure the result is expected in the input register.
Removed a TODO regarding the unnecessary usage of an FPU scope inside the crankshaft system. I created a similar todo at the top of the file because there are some other areas with the same "problem" that were unmarked.