[MIPS] Slight improvement to KeyedStoreStubCompiler::GenerateStoreFastDoubleElement()
The fix added to this function in afd90078 (returning input value_reg in v0) does not
affect test regress-91013.js, since the input value is always present in v0 also.

But it is fragile to depend on that behavior, so an explicit move from a0 to v0
is warranted. This slight improvement puts that move in the delay slot of the Ret().