[MIPS crankshaft] Detection and handling of the Erratum 156 problem in 24K cores.
This problem was discovered after commit 6675af80fe1433d48d6aa9cfbdc816b7c69c3730 and is caused by "jal" jumping to the next 256 MB segment instead of the current one in some specific cases.

This patch includes:
-a new routine in OS::MipsCpuHasFeature that detects the CPU type and revision and determines if it is affected by the bug.
-a special check in Assembler::set_target_address_at that makes sure this bug will not cause errors.

The handling for this bug is always disabled when running in the Simulator and always enabled when Serializing is enabled.