[MIPS crankshaft] Workaround for the 24K cores with the E156 bug.
Some 34k cores also have this bug, which is E168.

This workaround prevents the jalr-jal optimization if we are hitting
the 24K core E156 bug, which can occur when a jal instruction is within
the first or last 32K bytes of a 256MB page.

Since the excluded space is only 64KB out of 256MB (0.02%), we are
applying the workaround for all cores. That way we don't have to identify
the core type at run time, which has its own issues.

The workaround can be disabled by defining the symbol:
ALLOW_JAL_IN_BOUNDARY_REGION