[MIPS crankshaft] Initial MIPS crankshaft backport for v8 3.2 (ICS).
Note: this is by no means final. The generic system works but there are
still a lot of test failures, TODOs and areas where cleanup is necessary.

Most of the code found in this patch comes from the lithium code at
56bdbd0230827dfe60803f22d5bbe10823eb1678 (v8m-rb). Small parts of the code were
taken from the ARM version or manually backported to match the rest of
the code and interfaces in the 3.2 backport.