[MIPS merge] ARM: Implement DoInstanceOfKnownGlobal stub
Trial cherry-pick of f272bec, but this commit alone is not sufficient.
The crankshhaft implementation of InstanceOfKnownGlobal (with code-patching)
needs to be applied.

This is just a minor refactoring for MIPS so the code structure matches
the ARM version. The crankshaft-specific part of the commit has not been
implemented.

Ported commits:
14fa36adeb6581f22d235add882d3738f9b07b46