[MIPS] Fix LCodeGen::DoOuterContext() to use Context::CLOSURE_INDEX.
Loading contexts used Context::CLOSURE_INDEX at the time of the backport,
so make crankshaft use the same context as fcg and macro-asm.

This fixes the Earley-Boyer crash that was most easily seen in debug
mode, but occasionaly in release mode.

Note that commit 974586d (r8248) from June 9, 2011 changed all this
to use the method we had in our crankshaft. But the backport rev is
earlier.