[MIPS] Fix clearing of context in top frame.
This was a small porting error, we need to store zero, not a3 (r3),
which had been loaded with zero in the arm version.

I saw no evidence that this actually caused an error. Just noticed it.

This needs to be fixed in current bleeding_edge, and google's repo. It
still exists in both places.