[MIPS] Fix a BinaryOpStub marker bug.
This is Douglas's original fix, backported to the 3.2 branch.
Some of the patching stuff was a little more primitive here.

Still the critical fact is making the marker an andi to the
zero_reg, to avoid any possible conflict with normal emitted code.

This fixes mjsunit/array-sort.js and several mozilla tests in
debug mode.

Doug's original commit message:

This fixes a failure while running the negate.js test from mjsunit.
In crankshaft mode, a call to BinaryOpStub is not followed by a marker,
yet in PatchInlinedSmiCode(), the existence of a marker is checked in
full-codegen and crankshaft modes. Therefore if the marker is a normally
used instruction, it could lead to a false detection of the marker and
cause the program to access random locations in memory.

Note that google just commited 14752c97 to add a nop as the marker
in other arch. With this patch, we do not nbeed that nop.
(cherry picked from commit 5d87dac16ecbcf80c689550055a17eab5e9680fb)

Conflicts:

	src/mips/full-codegen-mips.cc (fixed).