Refactored iterative map traversal.
The main goal is to cleanly separate between the several parts involved in the traversal:

 * iterating over all transitions in a descriptor array

 * iterating over all prototype transitions

 * storing the parent and the current local traversal position in a map

 * the iterative traversal algorithm itself

The previous algorithm for iterating over prototype transitions did a little bit too much here, iterating over the whole array instead only the filled part. This has been fixed on the way, too.

With this CL, it will be much easier to make the necessary changes to the descriptor array iterator to correctly handle map transitions for accessor properties. Furthermore, perhaps we represent transitions a bit different in the future, making finding them a bit easier. This would make some code in this CL (and elsewhere) quite a bit shorter and more efficient.

Review URL: https://chromiumcodereview.appspot.com/9252007

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@10473 ce2b1a6d-e550-0410-aec6-3dcde31c8c00