MIPS: Reverted the maximal number of deopt entries to 4096.
On MIPS we cannot double the maximal number of deoptimization entries easily,
because the generated code would contain long branches (>128K) and we use
BlockTrampolinePoolScope in the Deoptimizer::TableEntryGenerator::GeneratePrologue method.
Also, we cannot use the Jr instruction here, because the Deoptimizer::CreateCode method asserts,
that the generated code does not contain reloc information.
One solution would be to use more, smaller branches, but this would cause performance degradation.