Merged r11831 into trunk branch.
Revert part of 11727 as it sometimes tanked V8 benchmark (raytrace) performance for reasons that are not obvious.  Now we make objects into fast-case objects when they are made prototypes for other objects, but we do not mark objects that are already fast case with a bit that helps keep them in fast case. Review URL: https://chromiumcodereview.appspot.com/10556004

R=danno@chromium.org
BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10564002

git-svn-id: https://v8.googlecode.com/svn/trunk@11832 ce2b1a6d-e550-0410-aec6-3dcde31c8c00